Change name for options.extensions to options.filter and change from an extension string to a RegExp.
Turns out that `options.extensions` for `express.static()` is not about which
files to handle but to add file-extensions to search patterns. E.i. when
`request.path` is index, then also search for index.htm and index.html.

I do not want that behavior but a way to configure express-mime-sniff,
to only handle _some_ files. So I am changing the name to reflect that.