Implemented "silence" and "fallthrough" options.
The "silence" option will stop middleware from printing to stderr on error.

The "fallthrough" option is modeled after serve-static's "fallthrough" option,
and tells the middleware to forward any error to express (like file not found),
if it is false. The default is true.

If "silence" is set to true and "fallthough" is set to false, then express
will print the error to stderr, cancelling out the effect. Therefore the
tests in middleware-options.spec.js creates new nodejs processes, so that the
main stderr does not get polluted with error messages.